Oppenings_Meta_Analyst/requirements.txt.txt
requests
pandas
seaborn
matplotlib
python-chess
io
time
csv
